Extract resolvers and mutations for order
